refactor buildQuoteTx to a helper function; modify test
