+++ packages/swapper/src/manager/SwapperManager.test.ts
@@ -2,8 +2,15 @@
 import { SwapperType } from '../api'
 import { ThorchainSwapper, ZrxSwapper } from '../swappers'
 import { SwapperManager } from './SwapperManager'
+import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
+import Web3 from 'web3'
 
 describe('SwapperManager', () => {
+  const zrxSwapperDeps = {
+    web3: {} as unknown as Web3,
+    adapterManager: {} as unknown as ChainAdapterManager
+  }
+
   describe('constructor', () => {
     it('should return an instance', () => {
       const manager = new SwapperManager()
@@ -12,6 +19,7 @@ describe('SwapperManager', () => {
   })
 
   describe('addSwapper', () => {
+
     it('should add swapper', () => {
       const manager = new SwapperManager()
       manager.addSwapper(SwapperType.Thorchain, new ThorchainSwapper())
@@ -19,10 +27,11 @@ describe('SwapperManager', () => {
     })
 
     it('should be chainable', async () => {
+
       const manager = new SwapperManager()
       manager
         .addSwapper(SwapperType.Thorchain, new ThorchainSwapper())
-        .addSwapper(SwapperType.Zrx, new ZrxSwapper())
+        .addSwapper(SwapperType.Zrx, new ZrxSwapper(zrxSwapperDeps))
       expect(manager.getSwapper(SwapperType.Zrx)).toBeInstanceOf(ZrxSwapper)
     })
 
@@ -31,7 +40,7 @@ describe('SwapperManager', () => {
       expect(() => {
         swapper
           .addSwapper(SwapperType.Thorchain, new ThorchainSwapper())
-          .addSwapper(SwapperType.Thorchain, new ZrxSwapper())
+          .addSwapper(SwapperType.Thorchain, new ZrxSwapper(zrxSwapperDeps))
       }).toThrow('already exists')
     })
   })
@@ -47,7 +56,7 @@ describe('SwapperManager', () => {
       const swapper = new SwapperManager()
       swapper
         .addSwapper(SwapperType.Thorchain, new ThorchainSwapper())
-        .addSwapper(SwapperType.Zrx, new ZrxSwapper())
+        .addSwapper(SwapperType.Zrx, new ZrxSwapper(zrxSwapperDeps))
 
       expect(swapper.getSwapper(SwapperType.Thorchain)).toBeInstanceOf(ThorchainSwapper)
       expect(swapper.getSwapper(SwapperType.Zrx)).toBeInstanceOf(ZrxSwapper)
